TheFrame/sidecar/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
samsungtvws[async,encrypted]>=2.6.0
pyyaml>=6.0
wakeonlan>=3.0.0
